Fix #100 Uploads fails when using AFNetworking 2.6.1
The behavior of AFHTTPRequestSerializer changed in AFNetworking 2.6.1
which causes the HTTPBody of requests to be discarded.  Specifically,
FPAPIClient.m was calling copying a request with AFHTTPRequestSerialized
twice.  The first time in `-[FPAPIClient POST:parameters:usingOperationQueue:success:failure:]`,
then again in the override in `HTTPRequestOperationWithRequest:Success:failure`.

Due to the change in behavior in AFNetworking (https://github.com/AFNetworking/AFNetworking/pull/2868/files)
if a request has an empty query string, the new request created by AFHTTPRequestSerializer will also have
an empty body, which is what was happening the second time that the request was copied
with AFHTTPRequestSerializer.

It seems that FPAPIClient was relying on undefined behavior of the AFHTTPRequestSerializer initializer;
namely that it would copy the HTTPBody if the HTTPBody was present.

Rather than try to get a fix made to AFNetworking, I think it's better to workaround the problem here
by removing the override for HTTPRequestOperationWithRequest since the function seems to serve no
purpose. The original commit
https://github.com/Ink/ios-picker/commit/72f6fa0affd0fd3b59f0321da7579eafee91d573
does not shed any light on what exactly it's trying to accomplish with the user agent.