Fix #108 FPSourceControllerDelegate delegate is called twice
The FPSourceControllerDelegate didPickMediaWithInfo delegate method is called twice
because +[PHImageManager requestImageForAsset:], when called asynchronously,
may call its completion handler more than once.  The fix is to to set synchronous=YES
to indicate that the completion block should be called exactly once.

This fix is scoped so that it will not affect the other callers of
asyncFetchAssetThumbnailFromPHAsset.